[Leo] Add locators.
As discussed, this actually introduces an ambiguity, because for example
`a.b/c()` could be interpreted, at parsing time, either as a division of `a.b`
by `c()` or as a call of the function located by `a.b/c` on no arguments. But we
are already planning to pick a different syntax for locators, so for now we let
this go.