[Aleo] Remove a layer in the grammar formulation.
This does not change the language defined by the grammar. It only optimizes a
few rules to avoid an extra group layer `( ... )`, which is presumably a
leftover from a time in which it was preceded by some whitespace and/or
comments. The presence or absence of the layer are equivalent for most purposes,
but they lead to slightly different CSTs.