[Leo] Add the unit type and expression.
The unit type is considered a primitive type, since its (only) value is atomic,
i.e. does not consist of aggregated sub-values. However, it is not a named type,
because it is denoted via the symbols `(` `)` and not via an identifier or a
keyword. These distinctions are reflected in the modified rules, which are more
than just addint a unit type.

The unit expression is added as a primary expression.

Note that there may be whitespace and comments between the opening and closing
parentheses of the unit type or expression. Even though it would be generally a
discouraged practice to actually put whitespace and comments there, it seems
appropriate to allow that at the grammar level. If we wanted to disallow that,
we would have to add `()` as a new symbol lexeme.

Partially related to the previous point, I considered whether `()` could be a
unit literal, but opted against it.