[Leo] Add the `private` modifier.
This is allowed in function parameters and output types of transition functions
and finalizer functions (not helper functions). The grammar captures the
restriction on helper function's output types, but not on helper function's
parameters. It could, if we differentiated the function parameters, but it may
not be worth the complication.