[Leo] Add inline assembly.
In order to keep lexing context-independent (and thus simpler, doable as a first
pass prior to parsing, as opposing to having the parser call the lexer to get
the next token), we add a few keywords that are only used in assembly blocks.

(We also swap two existing keywords, `inline` and `increment`, which were out of
order for some reason. This swap is unrelated to inline assembly.)

Note that the `caller` in `self.caller` is an identifier, not a keyword. This
seems more flexible, as we may add more such accessors, and there is no
syntactic ambiguity.

We allow negated literals as operands. This causes no issues in inline
instructions, which do not have expressions like `x-1` as in the rest of Leo,
which is the reason why Leo does not have negative literals (because otherwise
the longest lexeme rule would lex `x-1` as the identifier `x` followed by the
negative literal `-1`).

We allow operands that are non-empty sequences separated by `.`, but we do not
have a separate syntactic alternative for program IDs, which have the form
`<identifier>.<identifier>`, because that would create an ambiguity. This issue
is being discussed separately.

Consistently with our treatment of `operator-call`s in Leo, we do not list all
instructions, but instead we categorize them based on their "shape":
- 1 input to 1 output
- 2 inputs to 1 output
- 3 inputs to 1 output
- 2 inputs to 0 outputs
We use 'unary', 'binary', and 'ternary' for the first three, while the fourth
one is just for `assert`, and so we call it like that.

Note that, based on previous discussions, not all Aleo instructions have
counterparts here, by design.