[Leo] Generalize operator calls to method calls.
Instead of just allowing operator (method-like) calls for unary and binary
operators, the grammar now has general method calls. The underlying idea is that
an associated function with a `self` first parameter is also a method that can
be alternatively called with a dotted notation (which should be consistent with
Rust).